Address Archana's PR comments
