+++ sharktank/sharktank/models/t5/t5.py
@@ -26,78 +26,9 @@ from ...layers import (
 )
 from ... import ops
 from ...types.theta import Theta
-from ...layers import FFN
+from ...layers import FFN, T5Config
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class T5Config:
-    return_dict: bool = True
-    output_hidden_states: bool = False
-    output_attentions: bool = False
-    is_encoder_decoder: bool = True
-    is_decoder: bool = False
-    vocab_size: int = 32128
-    d_model: int = 512
-    d_kv: int = 64
-    d_ff: int = 2048
-    num_layers: int = 6
-    num_decoder_layers: int = 6
-    num_heads: int = 8
-    relative_attention_num_buckets: int = 32
-    relative_attention_max_distance: int = 128
-    layer_norm_epsilon: float = 1e-6
-    feed_forward_proj: str = "relu"
-    is_gated_act: bool = field(init=False)
-    activation_dtype: torch.dtype = torch.float32
-    dense_act_fn: str = field(init=False)
-    use_cache: bool = True
-    pad_token_id: int = 0
-    eos_token_id: int = 1
-    decoder_start_token_id: int = 0
-
-    def __post_init__(self):
-        self.is_gated_act = self.feed_forward_proj.startswith("gated-")
-        self.dense_act_fn = (
-            self.feed_forward_proj.split("-")[1]
-            if "-" in self.feed_forward_proj
-            else self.feed_forward_proj
-        )
-        if self.dense_act_fn == "gelu":
-            self.dense_act_fn = "gelu_new"
-
-    @staticmethod
-    def from_gguf_properties(properties: dict[str, Any], **kwargs):
-        assert properties["general.architecture"] == "t5"
-        assert (
-            properties["t5.attention.layer_norm_epsilon"]
-            == properties["t5.attention.layer_norm_rms_epsilon"]
-        )
-
-        gguf_to_config_names_map = {
-            "t5.embedding_length": ["d_model"],
-            "t5.feed_forward_length": ["d_ff"],
-            "t5.block_count": ["num_layers", "num_decoder_layers"],
-            "t5.attention.head_count": ["num_heads"],
-            "t5.attention.key_length": ["d_kv"],
-            "t5.attention.layer_norm_epsilon": ["layer_norm_epsilon"],
-            "t5.attention.relative_buckets_count": ["relative_attention_num_buckets"],
-            "t5.decoder_start_token_id": ["decoder_start_token_id"],
-            "tokenizer.ggml.eos_token_id": ["eos_token_id"],
-            "tokenizer.ggml.padding_token_id": ["pad_token_id"],
-        }
-        all_kwargs = {"vocab_size": None, "feed_forward_proj": None}
-        all_kwargs.update(
-            {
-                config_name: properties[gguf_name]
-                for gguf_name, config_names in gguf_to_config_names_map.items()
-                for config_name in config_names
-            }
-        )
-        all_kwargs.update(kwargs)
-
-        return T5Config(**all_kwargs)
 
 
 ACT2FN = {
